Completed basic wiring, beginning troubleshooting
Synchronous problems are starting to happen, where the code is hung up
on processing a single comment instead of letting it go on
asynchronously by itself. Also, I'm not altogether convinced that
placing a map in the state and constantly copying it and updating it
is the best idea. I need to look into this.