More short-term planning
I was originally planning to finish up CommentProcessor's functionality
today, but for some reason, SnooStream isn't working in that it's
not providing any comments. (I checked with my other comment script
and it's facing the same problem.)
In the meantime, I'm just planning out some work for maybe the next week
or so? The data displays are something that I'm looking forward to
working on.