Fixed mounting issue, stream continuity ensured
Fixed an issue where after the See Data button was pressed, even though
the CommentView component was hidden, it was still trying to update
with setState(). With this fix, the commentstream should hopefully
be ready to immediately get back to work after data has been visualized,
as the commentstream never really stops as a result of this fix.