Generalized button, grappling with reset button
I made the button more generalized of a prop, changing its name to
reflect its new status and also giving it the prop of className to more
easily edit its appearance. Also, working with reset button is annoying
mostly due to the fact that the way that React handles state changes
means that applying a class and then stripping that class currently
requires the use of setTimeout(), which obviously isnt the best
solution. Well, it's a work in progress, I guess.