Starting on Graphs
Besides a few minor changes (button appearance, box-shadow), this
commit is all about making the graph appear! I've finished the wiring
needed to provide the graph with data, but the graph's appearance
is still a work in progress. I might change the means by which I'm
implementing the graph, but this is a great learning experience!